Enlace de las coordenadas gráficas con los identificadores de las calles

+++ headers/hash_table_street_info.h
@@ -27,7 +27,7 @@ typedef struct TableStreetInfo {
 
 TableStreetInfo *create_table_street_info(int size);
 
-StreetInfo *create_street_info(int, int, StreetDir);
+StreetInfo *create_street_info(float, float, StreetDir);
 
 int hash_code_street_info(TableStreetInfo *, int);
